Add files via upload
same error as before, had to manually upload. i think the issue is that there are already files in the repository that have the same name/location and I can't overwrite with a new file without pulling the old one but i don't know how to do that yet and it's midnight. this file is the script for exercise 2